OS-7800. Rename rest to restapi
