--- ngui/ui/src/graphql/api/rest/index.ts
@@ -1,3 +1,0 @@
-import { GET_DATA_SOURCE } from "./rest.queries";
-
-export { GET_DATA_SOURCE };
