+++ ngui/ui/src/graphql/api/restapi/queries/index.ts
@@ -1,0 +1,3 @@
+import { GET_DATA_SOURCE } from "./restapi.queries";
+
+export { GET_DATA_SOURCE };
