+++ ngui/ui/src/graphql/api/restapi/queries/restapi.queries.ts
